Moved everything back to /app fixed package naming for android
