+++ app/android/app/src/debug/AndroidManifest.xml
