+++ app/android/app/src/main/AndroidManifest.xml
