+++ app/android/app/src/main/res/drawable/launch_background.xml
