+++ app/android/gradle.properties
