+++ app/android/settings.gradle
