--- pocket_therapist/android/app/src/main/kotlin/com/example/pocket_therapist/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.pocket_therapist
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}
